requirements.txt
wmi
termcolor
pyyaml